alpha-quant-v4/requirements.txt
# Alpha Quant v4.0 依赖清单

# 数据处理
pandas>=1.5.0
numpy>=1.23.0

# 数据获取
akshare>=1.10.0
tushare>=1.2.0
baostock>=0.8.8

# 机器学习
scikit-learn>=1.2.0

# 消息总线
redis>=4.5.0

# 任务调度
apscheduler>=3.10.0

# Web框架（信号服务器）
flask>=2.3.0
flask-cors>=4.0.0

# 日志
loguru>=0.7.0

# 环境变量
python-dotenv>=1.0.0

# HTTP请求
requests>=2.28.0

# 时间处理
pytz>=2023.3

# === V4.0 新增依赖 ===
# 机器学习
scikit-learn>=1.3.0
xgboost>=2.0.0

# 异步Web框架（替代Flask）
fastapi>=0.110.0
uvicorn>=0.25.0
websockets>=12.0

# Agent通信
redis>=5.0.0

# 自然语言处理
jieba>=0.42.1
